Edit a model

Edits the model in the passed project using the passed list. Fails if a model with the passed name already exists.

edit_model(zoltar_connection, model_url, model_config)

Arguments

zoltar_connection

A ZoltarConnection object as returned by new_connection()

model_url

url of a project in zoltar_connection's projects. this is the project the new model will be edited in

model_config

A list containing a Zoltar model configuration. An example: example-model-config.json . Full documentation at https://docs.zoltardata.com/.

Examples

if (FALSE) {
  edit_model(conn, "https://www.zoltardata.com/api/model/2/",
    jsonlite::read_json("example-model-config.json"))
}