Upload a forecast

This function submits forecast data to the server for uploading. Returns a Job object that can be used to up, which depends on the number of current uploads in the queue. Zoltar tracks these via Job objects.)

upload_forecast(
  zoltar_connection,
  model_url,
  timezero_date,
  forecast_data,
  is_json = TRUE,
  notes = ""
)

Arguments

zoltar_connection

A ZoltarConnection object as returned by new_connection()

model_url

URL of a model in zoltar_connection's projects

timezero_date

The date of the project timezero you are uploading for. it is a string in format YYYYMMDD

forecast_data

Forecast data in Zoltar format as a list or dataframe, depending on is_json

is_json

TRUE if forecast_data is JSON (list) format, and FALSE if it is CSV (dataframe) format

notes

Optional user notes for the new forecast

Value

A Job URL for the upload

Examples

if (FALSE) {
  forecast_data <- jsonlite::read_json("docs-predictions.json")
  job_url <- upload_forecast(conn, "http://www.zoltardata.com/api/model/1/",
                             "2017-01-17", forecast_data, TRUE, "a mid-January forecast")
}